refactor: move the event buffer into dial9-core
