+++ dial9-core/Cargo.toml
@@ -14,6 +14,7 @@ crossbeam-queue = "0.3"
 dial9-trace-format = { workspace = true }
 libc = { workspace = true }
 shuttle = { version = "0.9.1", optional = true }
+tracing = "0.1.44"
 
 [features]
 _shuttle = ["dep:shuttle"]
