+++ dial9-tokio-telemetry/src/telemetry/mod.rs
@@ -9,7 +9,7 @@ pub mod analysis;
 /// Decode-side companion structs for built-in trace events.
 #[cfg(any(feature = "analysis", test))]
 pub mod analysis_events;
-pub(crate) mod buffer;
+pub(crate) use dial9_core::buffer;
 pub(crate) use dial9_core::collector;
 pub use dial9_core::collector::Batch;
 #[cfg(feature = "cpu-profiling")]
@@ -26,8 +26,8 @@ pub(crate) mod task_metadata;
 pub(crate) mod writer;
 
 pub use crate::traced::TracedFuture;
-pub use buffer::{Encodable, ThreadLocalEncoder};
 pub use custom_events::{CustomEventsConfig, CustomEventsContext};
+pub use dial9_core::buffer::{Encodable, ThreadLocalEncoder};
 pub use events::{CpuSampleSource, clock_monotonic_ns};
 pub use format::{
     AllocEvent, FreeEvent, PollEndEvent, PollStartEvent, ProcessResourceUsageEvent, TaskSpawnEvent,
